Added filtering/sorting/agination using a queryobject

+++ api/Controllers/CommentController.cs
@@ -28,6 +28,9 @@ namespace api.Controllers
         [HttpGet("GetAllComments")]
         public async Task<IActionResult> GetAllComments()
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comments = await _commentRepo.GetAllCommentsAsync();
 
             var commentDto = comments.Select(s => s.ToCommentDto());
@@ -35,9 +38,12 @@ namespace api.Controllers
             return Ok(commentDto);
         }
 
-        [HttpGet("GetCommentById/{id}")]
+        [HttpGet("GetCommentById/{id:int}")]
         public async Task<IActionResult> GetCommentById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.GetCommentByIdAsync(id);
 
             if (comment == null)
@@ -48,39 +54,54 @@ namespace api.Controllers
             return Ok(comment.ToCommentDto());
         }
 
-        [HttpPost("CreateComment/{stockId}")]
+        [HttpPost("CreateComment/{stockId:int}")]
         public async Task<IActionResult> CreateComment([FromRoute] int stockId, CreateCommentDto commentDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             if (!await _stockRepo.StockExists(stockId))
             {
                 return BadRequest("Stock does not exist.");
             }
 
             var commentModel = commentDto.ToCommentFromCreate(stockId);
+
             await _commentRepo.CreateCommentAsync(commentModel);
-            return CreatedAtAction(nameof(GetCommentById), new { id = commentModel }, commentModel.ToCommentDto());
+
+            return CreatedAtAction(nameof(GetCommentById), new { id = commentModel.Id }, commentModel.ToCommentDto());
         }
 
-        [HttpPut("UpdateComment/{id}")]
+        [HttpPut("UpdateComment/{id:int}")]
         public async Task<IActionResult> UpdateComment([FromRoute] int id, [FromBody] UpdateCommentRequestDto updateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var comment = await _commentRepo.UpdateCommentAsync(id, updateDto.ToCommentFromUpdate());
+
             if (comment == null)
             {
                 return NotFound("Comment not found.");
             }
+
             return Ok(comment.ToCommentDto());
 
         }
 
-        [HttpDelete("DeleteComment/{id}")]
+        [HttpDelete("DeleteComment/{id:int}")]
         public async Task<IActionResult> DeleteComment([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var commentModel = await _commentRepo.DeleteCommentAsync(id);
+
             if (commentModel == null)
             {
                 return NotFound("Comment not found.");
             }
+
             return NoContent();
         }
