+++ api/Controllers/StockController.cs
@@ -4,6 +4,7 @@ using api.Mappers;
 using api.Dtos.Stock;
 using Microsoft.EntityFrameworkCore;
 using api.Interfaces;
+using api.Helpers;
 
 namespace api.Controllers
 {
@@ -20,18 +21,24 @@ namespace api.Controllers
         }
 
         [HttpGet("GetAllStocks")]
-        public async Task<IActionResult> GetAllStocks()
+        public async Task<IActionResult> GetAllStocks([FromQuery] QueryObject query)
         {
-            var stocks = await _stockRepo.GetAllStocksAsync();
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
+            var stocks = await _stockRepo.GetAllStocksAsync(query);
 
             var stockDto = stocks.Select(s => s.ToStockDto());
 
             return Ok(stocks);
         }
 
-        [HttpGet("GetStockById/{id}")]
+        [HttpGet("GetStockById/{id:int}")]
         public async Task<IActionResult> GetStockById([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stock = await _stockRepo.GetStockByIdAsync(id);
 
             if (stock == null)
@@ -45,14 +52,22 @@ namespace api.Controllers
         [HttpPost("CreateStock")]
         public async Task<IActionResult> CreateStock([FromBody] CreateStockRequestDto stockDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = stockDto.ToStockFromCreateDto();
+
             await _stockRepo.CreateStockAsync(stockModel);
+
             return CreatedAtAction(nameof(GetStockById), new { id = stockModel.Id }, stockModel.ToStockDto());
         }
 
-        [HttpPut("UpdateStock/{id}")]
+        [HttpPut("UpdateStock/{id:int}")]
         public async Task<IActionResult> UpdateStock([FromRoute] int id, [FromBody] UpdateStockRequestDto updateDto)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.UpdateStockAsync(id, updateDto);
 
             if (stockModel == null)
@@ -64,10 +79,14 @@ namespace api.Controllers
 
         }
 
-        [HttpDelete("DeleteStock/{id}")]
+        [HttpDelete("DeleteStock/{id:int}")]
         public async Task<IActionResult> DeleteStock([FromRoute] int id)
         {
+            if(!ModelState.IsValid)
+                return BadRequest(ModelState);
+
             var stockModel = await _stockRepo.DeleteStockAsync(id);
+
             if (stockModel == null)
             {
                 return NotFound();
